Init KDecoder (Mp4) .... Optimize...

+++ library/src/main/java/com/google/android/exoplayer2/decoder/KDecoder.java
@@ -71,9 +71,9 @@ public class KDecoder {
     /**
      * Origin file decode offset (Used to jump over do not need to decode)
      */
-    private int decodeOffset;
+    private long decodeOffset;
 
-    public KDecoder decodeOffset(int decodeOffset) {
+    public KDecoder decodeOffset(long decodeOffset) {
         this.decodeOffset = decodeOffset;
         return this;
     }
